rename: PeopleRow -> PeopleCell
