+++ app/[locale]/people/PeopleCell.tsx
